Moved the LoggingConfig into the Domain project and updated the Global.asax.cs. This is so it can be used in the other worker role projects
