+++ src/SFA.DAS.EmployerApprenticeshipsService.Domain/SFA.DAS.EmployerApprenticeshipsService.Domain.csproj
@@ -128,6 +128,7 @@
     <Compile Include="InvitationView.cs" />
     <Compile Include="LevyDeclarationView.cs" />
     <Compile Include="LevyItemType.cs" />
+    <Compile Include="Logging\LoggingConfig.cs" />
     <Compile Include="Membership.cs" />
     <Compile Include="Models\HmrcLevy\Declarations.cs" />
     <Compile Include="Models\HmrcLevy\EnglishFractionDeclarations.cs" />
