+++ src/SFA.DAS.EmployerApprenticeshipsService.Web/SFA.DAS.EmployerApprenticeshipsService.Web.csproj
@@ -262,7 +262,6 @@
   <ItemGroup>
     <Compile Include="App_Start\BundleConfig.cs" />
     <Compile Include="App_Start\FilterConfig.cs" />
-    <Compile Include="App_Start\LoggingConfig.cs" />
     <Compile Include="App_Start\RouteConfig.cs" />
     <Compile Include="App_Start\StructuremapMvc.cs" />
     <Compile Include="Authentication\IOwinWrapper.cs" />
